peptone_fit_model/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
openpyxl>=3.1.0  # For Excel file handling

# Scientific computing
scikit-learn>=1.3.0

# Bioinformatics
biopython>=1.81

# API and web
requests>=2.31.0

# UI (Phase 4)
streamlit>=1.28.0
plotly>=5.17.0
kaleido>=0.2.1  # For static image export

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
